Move PointSequence to an individual module
